Move Beatifyd data classes to dev.teogor.stitch package and rename AppDatabase to BeatifydDatabase
