+++ app/src/main/kotlin/dev/teogor/stitch/core/data/dao/HistoryDao.kt
@@ -14,12 +14,12 @@
  * limitations under the License.
  */
 
-package com.zeoowl.beatifyd.core.data.dao
+package dev.teogor.stitch.core.data.dao
 
 import androidx.room3.Dao
 import androidx.room3.Query
 import androidx.room3.Upsert
-import com.zeoowl.beatifyd.core.data.model.History
+import dev.teogor.stitch.core.data.model.History
 import dev.teogor.stitch.ExplicitEntities
 import kotlinx.coroutines.flow.Flow
